backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0

SQLAlchemy==2.0.29

python-dotenv==1.0.1

requests==2.31.0

# Optional (only if using OpenAI instead of Ollama)
openai==1.30.1